+++ src/components/plan/updatePlan/Pins.tsx
@@ -3,7 +3,7 @@ import { DndProvider } from 'react-dnd';
 import { HTML5Backend } from 'react-dnd-html5-backend';
 import { useParams } from 'react-router-dom';
 
-import { getCost, insertPlanEnding } from '@api/datesPay';
+import { getCost, insertPlanEnding } from '@api/datesPay';?
 import { calcPath } from '@api/path';
 import { type PinContentsType, getPin, deletePin } from '@api/pins';
 import IconPin from '@assets/icons/IconPin';
@@ -90,33 +90,33 @@ const Pins = ({ currentPage, dates }: PropsType) => {
     }
   }, [pinArr]);
 
-  // 08-30
-  const calcCostAndInsertPlansEnding = async () => {
-    const response = await getCost(planId);
+  // // 08-30
+  // const calcCostAndInsertPlansEnding = async () => {
+  //   const response = await getCost(planId);
 
-    if (response !== null && response !== undefined) {
-      const datesCost: number[] = [];
+  //   if (response !== null && response !== undefined) {
+  //     const datesCost: number[] = [];
 
-      response.forEach((value) => {
-        let cost = 0;
+  //     response.forEach((value) => {
+  //       let cost = 0;
 
-        value.contents.forEach((content) => {
-          cost += content.cost;
-        });
+  //       value.contents.forEach((content) => {
+  //         cost += content.cost;
+  //       });
 
-        datesCost.push(cost);
-      });
+  //       datesCost.push(cost);
+  //     });
 
-      console.log('result: ', datesCost);
-      console.log('distanceData: ', distanceData);
+  //     console.log('result: ', datesCost);
+  //     console.log('distanceData: ', distanceData);
 
-      void insertPlanEnding({
-        id: planId,
-        distance: distanceData.map(Number),
-        dates_cost: datesCost,
-      });
-    }
-  };
+  //     void insertPlanEnding({
+  //       id: planId,
+  //       distance: distanceData.map(Number),
+  //       dates_cost: datesCost,
+  //     });
+  //   }
+  // };
 
   return (
     <>
@@ -150,7 +150,7 @@ const Pins = ({ currentPage, dates }: PropsType) => {
       </button>
       <button
         className="p-5 bg-slate-500"
-        onClick={calcCostAndInsertPlansEnding}
+        // onClick={calcCostAndInsertPlansEnding}
       >
         여행 완료
       </button>
